Commit bugfix to interp_nbrhd() function.  I was doing integer division when is should be floating point.
Former-commit-id: 7fa5bb7ef1825bee3a3da9fb12ad236997a7da93

+++ met/src/basic/vx_util/interp_util.cc
@@ -334,7 +334,7 @@ double interp_nbrhd(const DataPlane &dp, int x_ll, int y_ll, int wdth,
       nbrhd_cov = bad_data_double;
    }
    else {
-      nbrhd_cov = count_thr/count;
+      nbrhd_cov = (double) count_thr/count;
    }
 
    return(nbrhd_cov);
